+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 
 ### Added
 
-* Added Helix endpoint `Manage Held AutoMod Messages`
+* Added Helix endpoints:
+  *  `Manage Held AutoMod Messages`
+  *  `Get Global Chat Badges`
+  *  `Get Channel Chat Badges`
 * Added Channel Terms to pubsub `chat_moderator_actions`
 * Added `user-moderation-notifications` topic to pubsub
 * Added `extendsub` to pubsub `channel-subscribe-events-v1`